Type updates to SBP spec.
Updates from piksi_firmware HITL integration with libsbp C headers:
0. Carrier phase struct is actually s32 and not u32.
1. Make UART field name informative.
2. Remove old sbp_messages.h

Closes #44.

/cc @mfine @fnoble
