+++ python/sbp/tracking.py
@@ -12,10 +12,9 @@
 from construct import *
 from sbp import SBP
 from sbp.utils import fmt_repr
-from sbp.lib import *
 
 # Automatically generated from piksi/yaml/swiftnav/sbp/tracking.yaml
-# with generate.py at 2015-03-21 14:14:48.340628. Please do not hand edit!
+# with generate.py at 2015-03-23 14:30:01.621847. Please do not hand edit!
 
 
 class TrackingChannelState(object):
