--- spec/yaml/swiftnav/sbp/lib.yaml
@@ -1,56 +1,0 @@
-# Copyright (C) 2015 Swift Navigation Inc.
-# Contact: Fergus Noble <fergus@swiftnav.com>
-#
-# This source is subject to the license found in the file 'LICENSE' which must
-# be be distributed together with this source. All other rights reserved.
-#
-# THIS CODE AND INFORMATION IS PROVIDED "AS IS" WITHOUT WARRANTY OF ANY KIND,
-# EITHER EXPRESSED OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE IMPLIED
-# WARRANTIES OF MERCHANTABILITY AND/OR FITNESS FOR A PARTICULAR PURPOSE.
-
-package: swiftnav.sbp.lib
-description: Defines a variety of common structs.
-stable: False
-include:
-  - types.yaml
-definitions:
-
- - Latency:
-    desc: |
-      Statistics on the latency of observations received from the base
-      station. As observation packets are received their GPS time is
-      compared to the current GPS time calculated locally by the
-      receiver to give a precise measurement of the end-to-end
-      communication latency in the system.
-    fields:
-        - avg:
-            type: s32
-            units: ms
-            desc: Average latency.
-        - lmin:
-            type: s32
-            units: ms
-            desc: Minimum latency.
-        - lmax:
-            type: s32
-            units: ms
-            desc: Maximum latency.
-        - current:
-            type: s32
-            units: ms
-            desc: Smoothed estimate of the current latency.
-
- - CarrierPhase:
-    desc: |
-      Carrier phase measurement in cycles represented as a 40-bit
-      fixed point number with Q32.8 layout, i.e. 32-bits of whole
-      cycles and 8-bits of fractional cycles.
-    fields:
-        - i:
-            type: u32
-            units: cycles
-            desc: Carrier phase whole cycles.
-        - f:
-            type: u8
-            units: cycles / 255
-            desc: Carrier phase fractional part.
